requirements.txt
asgiref==3.9.1
asttokens==3.0.0
decorator==5.2.1
django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
executing==2.2.0
gunicorn==23.0.0
ipdb==0.13.13
ipython==9.4.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.51
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
pyjwt==2.9.0
python-dotenv==1.1.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
